chore: address code-review nits (redundant int cast, dead test-req lines)
TimeDriftSensor: drop the redundant int() around round() (round() with no ndigits already returns int) and the RUF046 noqa that was suppressing it. requirements.test.txt: remove the loose pytest/pytest-cov/aiohttp lines that phacc already pins transitively. No behavior change; suite 76 passed, ruff clean.

+++ custom_components/eveus/sensor.py
@@ -266,7 +266,7 @@ class TimeDriftSensor(ChargerSensor):
         # Anti-flicker: in-sync clocks read exactly 0, real drift in 10s steps
         if abs(drift) < 30:
             return 0
-        return int(round(drift / 10.0)) * 10  # noqa: RUF046  explicit int() documents intent
+        return round(drift / 10.0) * 10
 
 
 class SessionEnergySensor(ChargerSensor, RestoreEntity):
